% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy.R
\name{greedy}
\alias{greedy}
\title{Greedy Registration with Affine}
\usage{
greedy(fixed, moving, affine = TRUE, dimension = 3,
  args.affine = list(niters = "200x200x200", omat = tempfile(fileext =
  ".mat"), metric = "NMI", dof = 12, opts = ""),
  args.greedy = list(initial_transform = args.affine$omat, niters =
  args.affine$niters, metric = args.affine$metric, warpfile = tempfile(fileext =
  ".nii.gz"), invwarpfile = tempfile(fileext = ".nii.gz"), opts = ""),
  verbose = TRUE, ...)
}
\arguments{
\item{fixed}{Fixed image}

\item{moving}{Moving Image to register to fixed image}

\item{affine}{Should an affine registration be run before greedy?}

\item{dimension}{Number of image dimension}

\item{args.affine}{list of arguments for affine registration and
\code{\link{greedy_affine}}}

\item{args.greedy}{list of arguments for greedy registration and
\code{\link{greedy_estimate}}}

\item{verbose}{print diagnostic messages}

\item{...}{additional arguments passed to \code{\link{greedy_apply}},
like \code{interpolator}}
}
\value{
List of output from all registrations and application
}
\description{
Registration from Greedy library
}
