% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_affine.R
\name{greedy_affine}
\alias{greedy_affine}
\title{Affine Registration from Greedy}
\usage{
greedy_affine(fixed, moving, niters = "100x50x10", dimension = 3,
  omat = tempfile(fileext = ".mat"), dof = 12, metric = "NCC 2x2x2",
  opts = "", verbose = TRUE)
}
\arguments{
\item{fixed}{Fixed image}

\item{moving}{Moving Image to register to fixed image}

\item{niters}{number of iterations per level of multi-res}

\item{dimension}{Number of image dimension}

\item{omat}{Output affine matrix}

\item{dof}{degrees of freedom}

\item{metric}{Metric to use for registraation}

\item{opts}{additional options}

\item{verbose}{print diagnostic messages}
}
\value{
List of output
}
\description{
Affine Registration from Greedy library
}
