% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_apply.R
\name{greedy_apply}
\alias{greedy_apply}
\title{Apply Transformations from Greedy}
\usage{
greedy_apply(fixed, moving, transforms = list(), moving_outfile = NULL,
  interpolator = "NN", warpfile = tempfile(fileext = ".nii.gz"),
  jacobian = tempfile(fileext = ".nii.gz"), dimension = 3, opts = "",
  verbose = TRUE)
}
\arguments{
\item{fixed}{Fixed image}

\item{moving}{Moving Image(s) to apply transformations}

\item{transforms}{List of transformations}

\item{moving_outfile}{List of output filenames corresponding to the moving
files}

\item{interpolator}{interpolation operation to use (NN, LINEAR*, LABEL sigma)}

\item{warpfile}{Output composed warp field filename}

\item{jacobian}{Output Jacobian determinant image}

\item{dimension}{Number of image dimension}

\item{opts}{additional options}

\item{verbose}{print diagnostic messages}
}
\value{
List of output warps and inverses
}
\description{
Apply Registration from Greedy library
}
