% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_estimate.R
\name{greedy_estimate}
\alias{greedy_estimate}
\title{Deformation Registration from Greedy}
\usage{
greedy_estimate(fixed, moving, initial_transform = NULL,
  warpfile = tempfile(fileext = ".nii.gz"), invwarpfile = tempfile(fileext =
  ".nii.gz"), niters = "100x50x10", dimension = 3, metric = "NCC 2x2x2",
  opts = "", verbose = TRUE)
}
\arguments{
\item{fixed}{Fixed image}

\item{moving}{Moving Image to register to fixed image}

\item{initial_transform}{Mat file of affine intialization}

\item{warpfile}{Output warp field filename}

\item{invwarpfile}{Inverse of the warp field filename}

\item{niters}{number of iterations per level of multi-res}

\item{dimension}{Number of image dimension}

\item{metric}{Metric to use for registraation}

\item{opts}{additional options}

\item{verbose}{print diagnostic messages}
}
\value{
List of output warps and inverses
}
\description{
Deformation Registration from Greedy library
}
